+++ user-management/backend/app/user-service/routes/routes.go
@@ -10,6 +10,11 @@ func Setup(app *iris.Application, publicUserService service.PublicUserService, p
 
 	gatewayController := controller.NewServiceControllerImp(publicUserService, privateUserService)
 	// auth routes
+	app.Post("/api/auth/login", gatewayController.Login())
+	app.Post("/api/login", gatewayController.Login())
+	app.Put("/api/login", gatewayController.Login())
+	app.Post("/api/auth/logout", gatewayController.Logout())
+	app.Put("/api/logout", gatewayController.Logout())
 	app.Post("/api/register", gatewayController.Register())
 	app.Get("/api/users/me", gatewayController.User())
 	app.Put("/api/users/info", gatewayController.UpdateInfo())